fix: remove -rlink-path arg
This allows elfutils to be compiled with clang, to be less dependent on
what toolchain the end user is using.

It seems that `-rlink-path` is a GNU ld arg only, so LLVM's `lld` will
fail if it encounters it. The library seems to build without, so I
suppose we could try it.